requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
streamlit==1.28.2
ultralytics==8.4.21
opencv-python-headless==4.8.1.78
Pillow==10.1.0
numpy==1.26.2
matplotlib==3.8.2